+++ apps/client/src/components/Footer/Footer.test.js
@@ -1,32 +1,18 @@
 import React from "react";
-import { MemoryRouter } from "react-router-dom";
 
-import Context from "../../__mocks__/context";
 import matchMedia from "../../__mocks__/matchMedia";
-import { topics } from "../../config";
 import { FOOTER } from "../../utils/test-ids";
 import { render } from "../../utils/test-utils";
 import Footer from ".";
 
 Object.defineProperty(window, "matchMedia", matchMedia);
 
-const Wrapper = ({ children }) => {
-  const topicMock = "dakraam-plaatsen";
-  const topicUrlMock = `/${topicMock}`;
-  const topic = topics.find((t) => t.slug === topicMock);
-
-  return (
-    <Context topicMock={topic}>
-      <MemoryRouter initialEntries={[topicUrlMock]}>{children}</MemoryRouter>
-    </Context>
-  );
-};
+jest.mock("react-router-dom", () => ({
+  ...jest.requireActual("react-router-dom"),
+  useParams: () => ({}),
+}));
 
 it("renders with text", () => {
-  const { getByTestId } = render(
-    <Wrapper>
-      <Footer />
-    </Wrapper>
-  );
+  const { getByTestId } = render(<Footer />);
   expect(getByTestId(FOOTER)).toBeTruthy();
 });