+++ apps/client/src/config/index.ts
@@ -1,8 +1,9 @@
 export type Topic = {
+  intro?: string;
+  name: string;
+  redirectToOlo?: boolean;
   slug: string;
   sttrFile?: string;
-  redirectToOlo?: boolean;
-  intro?: string;
   text: Text;
 };
 
@@ -48,49 +49,56 @@ export const generateOloUrl = ({
 
 const topics: Topic[] = [
   {
-    slug: "kappen-of-snoeien",
+    name: "Kappen of snoeien",
     redirectToOlo: true,
+    slug: "kappen-of-snoeien",
     text: {
       heading: "Vergunningcheck kappen of snoeien",
     },
   },
   {
+    intro: "DakkapelIntro",
+    name: "Dakkapel plaatsen",
     slug: "dakkapel-plaatsen",
     sttrFile: "dakkapel.json",
     text: {
       heading: "Vergunningcheck dakkapel plaatsen",
       locationIntro: "Voer het adres in waar u de dakkapel wilt gaan plaatsen",
     },
-    intro: "DakkapelIntro",
   },
   {
+    intro: "DakraamIntro",
+    name: "Dakraam plaatsen",
     slug: "dakraam-plaatsen",
     sttrFile: "dakraam.json",
     text: {
       heading: "Vergunningcheck dakraam plaatsen",
       locationIntro: "Voer het adres in waar u het dakraam wilt gaan plaatsen",
     },
-    intro: "DakraamIntro",
   },
   {
+    intro: "AanbouwIntro",
+    name: "Aanbouw of uitbouw maken",
     slug: "aanbouw-of-uitbouw-maken",
     text: {
       heading: "Vergunningcheck aanbouw of uitbouw maken",
       locationIntro:
         "Voer het adres in waar u de aanbouw of uitbouw wilt gaan maken",
     },
-    intro: "AanbouwIntro",
   },
   {
+    intro: "KozijnenIntro",
+    name: "Kozijnen plaatsen",
     slug: "kozijnen-plaatsen",
     sttrFile: "kozijn.json",
     text: {
       heading: "Vergunningcheck kozijnen plaatsen",
       locationIntro: "Voer het adres in waar u de kozijnen wilt gaan plaatsen",
     },
-    intro: "KozijnenIntro",
   },
   {
+    intro: "ZonnepanelenIntro",
+    name: "Zonnepanelen of zonneboiler plaatsen",
     slug: "zonnepanelen-of-zonneboiler-plaatsen",
     sttrFile: "zonnepaneel.json",
     text: {
@@ -98,25 +106,28 @@ const topics: Topic[] = [
       locationIntro:
         "Voer het adres in waar u de zonnepanelen of zonneboiler wilt gaan plaatsen",
     },
-    intro: "ZonnepanelenIntro",
   },
   {
+    intro: "BouwwerkSlopenIntro",
+    name: "Bouwwerk slopen",
     slug: "bouwwerk-slopen",
     text: {
       heading: "Vergunningcheck bouwwerk slopen",
       locationIntro: "Voer het adres in waar u het bouwwerk wilt gaan slopen",
     },
-    intro: "BouwwerkSlopenIntro",
   },
   {
+    intro: "InternVerbouwenIntro",
+    name: "Intern verbouwen",
     slug: "intern-verbouwen",
     text: {
       heading: "Vergunningcheck intern verbouwen",
       locationIntro: "Voer het adres in waar u intern wilt gaan verbouwen",
     },
-    intro: "InternVerbouwenIntro",
   },
   {
+    intro: "ZonweringRolluikIntro",
+    name: "Zonwering of rolluik plaatsen",
     slug: "zonwering-of-rolluik-plaatsen",
     sttrFile: "zonwering.json",
     text: {
@@ -124,19 +135,19 @@ const topics: Topic[] = [
       locationIntro:
         "Voer het adres in waar u de zonwering, het rolhek, rolluik of luik wilt gaan plaatsen",
     },
-    intro: "ZonweringRolluikIntro",
   },
 ];
 
 if (process.env.NODE_ENV !== "production") {
   topics.push({
+    intro: "DebugIntro",
+    name: "Test outcomes",
     slug: "test-outcomes",
     sttrFile: "outcomes.json",
     text: {
       heading: "Testing different outcomes/conclusions",
       locationIntro: "Pick a random address...",
     },
-    intro: "DebugIntro",
   });
 }