+++ apps/client/src/hooks/index.ts
@@ -1,7 +1,5 @@
 import useChecker from "./useChecker";
 import useTopic from "./useTopic";
+import useTracking from "./useTracking";
 
-export {
-  useChecker,
-  useTopic,
-};
+export { useChecker, useTopic, useTracking };