+++ apps/client/src/pages/IntroPage.tsx
@@ -5,7 +5,8 @@ import { useHistory } from "react-router-dom";
 import Layout from "../components/Layouts/DefaultLayout";
 import Loading from "../components/Loading";
 import Nav from "../components/Nav";
-import { useTopic } from "../hooks";
+import { actions, sections } from "../config/matomo";
+import { useTopic, useTracking } from "../hooks";
 import { geturl, routes } from "../routes";
 import NotFoundPage from "./NotFoundPage";
 import RedirectPage from "./RedirectPage";
@@ -13,6 +14,7 @@ import RedirectPage from "./RedirectPage";
 const IntroPage = () => {
   const history = useHistory();
   const topic = useTopic();
+  const { matomoTrackEvent } = useTracking();
 
   if (!topic) {
     return <NotFoundPage />;
@@ -25,6 +27,16 @@ const IntroPage = () => {
 
   const Intro = React.lazy(() => import(`../intros/${intro}`));
 
+  const goToNext = () => {
+    history.push(geturl(routes.checker, topic));
+
+    // @TODO: Change and refactor this because the next step is not always LOCATION_INPUT
+    matomoTrackEvent({
+      action: actions.ACTIVE_STEP,
+      name: sections.LOCATION_INPUT,
+    });
+  };
+
   return (
     <Layout>
       <Helmet>
@@ -33,11 +45,7 @@ const IntroPage = () => {
       <Suspense fallback={<Loading />}>
         <Intro />
       </Suspense>
-      <Nav
-        noMarginBottom
-        onGoToNext={() => history.push(geturl(routes.checker, topic))}
-        showNext
-      />
+      <Nav noMarginBottom onGoToNext={goToNext} showNext />
     </Layout>
   );
 };